requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1.10
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.0.1
openai==0.27.8
protobuf==4.21.12
PyMySQL==1.1.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
